+++ packages/sdk/src/global_config.js
@@ -61,7 +61,7 @@ function validatePath(
  *
  * The contents will be synced in real-time to all logged-in users of the installation.
  * Contents will not be updated in real-time when the installation is running in
- * a publicly shared base, or in development mode.
+ * a publicly shared base.
  *
  * Any key can be watched to know when the value of the key changes.
  *