+++ packages/sdk/src/ui/select.js
@@ -92,7 +92,7 @@ class Select extends React.Component<SelectProps> {
         this._select.click();
     }
     render() {
-        const {id, className, style, options: originalOptions = [], value} = this.props;
+        const {id, className, style, options: originalOptions = [], value, disabled} = this.props;
 
         const validationResult = validateOptions(originalOptions);
         if (!validationResult.isValid) {
@@ -127,8 +127,8 @@ class Select extends React.Component<SelectProps> {
                 className={classNames(
                     'styled-input p1 rounded normal no-outline darken1 text-dark',
                     {
-                        'link-quiet pointer': !this.props.disabled,
-                        quieter: this.props.disabled,
+                        'link-quiet pointer': !disabled,
+                        quieter: disabled,
                     },
                     className,
                 )}
@@ -138,6 +138,7 @@ class Select extends React.Component<SelectProps> {
                 }}
                 value={optionValueToString(value)}
                 onChange={this._onChange}
+                disabled={disabled}
             >
                 {options &&
                     options.map(option => {