+++ packages/sdk/src/ui/select_buttons.js
@@ -1,6 +1,7 @@
 // @flow
 import classNames from 'classnames';
 import * as React from 'react';
+import omit from 'lodash.omit';
 import {spawnError} from '../error_utils';
 
 /* eslint-disable react/prop-types */
@@ -12,7 +13,6 @@ import {
     type SelectAndSelectButtonsProps,
 } from './select_and_select_buttons_helpers';
 
-const u = window.__requirePrivateModuleFromAirtable('client_server_shared/u');
 const KeyCodes = window.__requirePrivateModuleFromAirtable('client_server_shared/key_codes');
 
 /** @typedef */
@@ -43,7 +43,7 @@ class SelectButtons extends React.Component<SelectButtonsProps> {
             throw spawnError('<SelectButtons> %s', validationResult.reason);
         }
 
-        const restOfProps = u.omit(this.props, Object.keys(SelectButtons.propTypes));
+        const restOfProps = omit(this.props, Object.keys(SelectButtons.propTypes));
 
         return (
             <div