+++ packages/sdk/src/ui/select_buttons_synced.js
@@ -1,13 +1,12 @@
 // @flow
 import * as React from 'react';
+import omit from 'lodash.omit';
 import SelectButtons from './select_buttons';
 import {
     SelectAndSelectButtonsSyncedPropTypes,
     type SelectAndSelectButtonsSyncedProps,
 } from './select_and_select_buttons_helpers';
 import Synced from './synced';
-
-const {u} = window.__requirePrivateModuleFromAirtable('client_server_shared/hu');
 
 /** @typedef */
 type SelectButtonsSyncedProps = SelectAndSelectButtonsSyncedProps;
@@ -17,7 +16,7 @@ class SelectButtonsSynced extends React.Component<SelectButtonsSyncedProps> {
     static propTypes = SelectAndSelectButtonsSyncedPropTypes;
     props: SelectButtonsSyncedProps;
     render() {
-        const restOfProps = u.omit(this.props, ['globalConfigKey', 'onChange', 'disabled']);
+        const restOfProps = omit(this.props, ['globalConfigKey', 'onChange', 'disabled']);
         return (
             <Synced
                 globalConfigKey={this.props.globalConfigKey}